+++ metadata.rb
@@ -1,10 +1,10 @@
 name             'qmailtoaster'
 maintainer       'Sebastian Grewe'
 maintainer_email 'sebastian@grewe.ca'
-license          'All rights reserved'
+license          'Apache 2.0'
 description      'Installs/Configures qmailtoaster and Spamdyke (optional)'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.0.3'
+version          '0.0.4'
 
 recipe           'qmailtoaster', 'Installs and configures a qmail toaster installation, includes many sub-recipes'
 recipe           'qmailtoaster::_add_repositories', 'Included by default: Adds required RPMForge Repository'