+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 This file is used to list changes made in each version of qmailtoaster.
 
+## 0.0.4:
+
+* Changed license to Apache 2.0
+* Added Kitchen 1.0.0 Alpha Support
+* Added various minitests to check the installation
+* Updated qmailadmin-toaster source RPM version
+* Install build-essential prior to anything else
+* Fixed issue with submission running out of memory
+* Fixed issue with pre setup not removing packages and users
+* Added [codeship.io](http://www.codeship.io) for CI tests
+* Cleaned up [README.md](README.md)
+
+## 0.0.3:
+
+## 0.0.2:
+
 ## 0.0.1:
 
 * Initial release of qmailtoaster